% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_handling_functions.R
\name{get_diagnostics}
\alias{get_diagnostics}
\title{Get diagnostics from position data}
\usage{
get_diagnostics(pos_data, acc_data = NULL)
}
\arguments{
\item{pos_data}{A data frame containing position data with a POSIXct date_time column
and position measurements OR an empty dataframe.}

\item{acc_data}{A data frame containing acceleration data, used as a fallback if an empt pos_data is provided}
}
\value{
A data frame containing diagnostics information such as deployment length,
number of positions, and statistics on satellite counts and accuracy. If an empty dataframe is provided, this will largely be NA.
}
\description{
Get diagnostics from position data
}
\examples{
\dontrun{
get_diagnostics(pos_data)
}
}
