% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_handling_functions.R
\name{get_tag_files}
\alias{get_tag_files}
\title{Get all files for a specific tag}
\usage{
get_tag_files(tag_id, target_file_path = NULL, file_paths = NULL)
}
\arguments{
\item{tag_id}{The ID of the tag to retrieve files for.}

\item{target_file_path}{A path to the directory containing the tag files.
Alternatively, you can provide a vector of file names directly using the \code{file_paths} parameter.}

\item{file_paths}{A vector of file names to search for the specified tag ID.}
}
\value{
A vector of file paths associated with the specified tag ID.
}
\description{
Get all files for a specific tag
}
\examples{
get_tag_files("61029", "path/to/directory")
}
