% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_handling_functions.R
\name{load_all_gps_data}
\alias{load_all_gps_data}
\title{Read all data from an archive or folder}
\usage{
load_all_gps_data(folder_path = NULL, archive_path = NULL)
}
\arguments{
\item{folder_path}{The path to the folder containing the tag data files (if applicable).}

\item{archive_path}{The path to the archive file (if applicable).}
}
\value{
A list of lists, each containing data frames for a specific tag:
\itemize{
\item acc_immersion_data: A data frame containing acceleration data with immersion information.
\item acc_data: A data frame containing acceleration data without immersion information.
\item pos_data: A data frame containing position data.
\item diag_data: A data frame containing diagnostics information.
\item root_folders: vector of filepaths from which tag data were loaded.
\item base_path: folder_path or archive_path which was used to load files.
}
}
\description{
Given a folder path or an archive path, this function reads all tag data files
and returns a list of data frames for each tag, including acceleration data (with and
without immersion), position data, and diagnostics.
}
\examples{
\dontrun{
load_all_data(folder_path = "path/to/directory") # Load data from a folder
load_all_data(archive_path = "path/to/archive.zip") # Load data from an archive
}
}
