% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_handling_functions.R
\name{load_tag_acc_data}
\alias{load_tag_acc_data}
\title{Load acceleration data for a specific tag}
\usage{
load_tag_acc_data(tag_id, tag_files, archive_path = NULL, immersion = TRUE)
}
\arguments{
\item{tag_id}{The ID of the tag to retrieve acceleration data for.}

\item{tag_files}{character vector of file paths associated with the specified tag ID.}

\item{archive_path}{The path to the archive file (if applicable).}

\item{immersion}{Logical indicating whether to read immersion data (default is TRUE).}
}
\value{
A data frame containing the acceleration data with a POSIXct date_time column
and acceleration measurements.
}
\description{
Load acceleration data for a specific tag
}
\examples{
\dontrun{
load_tag_acc_data("61029", "path/to/archive", immersion = TRUE)
}
}
