% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_handling_functions.R
\name{load_tag_pos_data}
\alias{load_tag_pos_data}
\title{Load position data for a specific tag}
\usage{
load_tag_pos_data(tag_id, tag_files, archive_path = NULL)
}
\arguments{
\item{tag_id}{The ID of the tag to retrieve position data for.}

\item{tag_files}{character vector of file paths associated with the specified tag ID.}

\item{archive_path}{The path to the archive file (if applicable).}
}
\value{
A data frame containing the position data with a POSIXct date_time column
}
\description{
Load position data for a specific tag
}
