% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_handling_functions.R
\name{read_acc_data}
\alias{read_acc_data}
\title{Get acceleration data from a file connection}
\usage{
read_acc_data(tag_id, file_connection, immersion = TRUE)
}
\arguments{
\item{tag_id}{The ID of the tag to retrieve acceleration data for.}

\item{file_connection}{A connection to the file containing the acceleration data.}

\item{immersion}{Logical indicating whether to read immersion data (default is TRUE).}
}
\value{
A data frame containing the acceleration data with a POSIXct date_time column
and acceleration measurements.
}
\description{
Get acceleration data from a file connection
}
\examples{
get_acc_data("61029", "path/to/directory", immersion = TRUE)
}
