% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_handling_functions.R
\name{read_pos_data}
\alias{read_pos_data}
\title{Get position data from a specific tag file}
\usage{
read_pos_data(tag_id, file_connection, sep = ",", dec = ".")
}
\arguments{
\item{tag_id}{The ID of the tag to retrieve position data for.}

\item{file_connection}{FILE CONNECTION}

\item{sep}{The separator used in the data file (default is comma).}

\item{dec}{The decimal point character used in the data file (default is period).}
}
\value{
A data frame containing the position data with a POSIXct date_time column
and position measurements.
}
\description{
Get position data from a specific tag file
}
\examples{
get_pos_data("61029", "path/to/directory")
}
