% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seatrackgps-package.R
\docType{package}
\name{seatrackgps-package}
\alias{seatrackgps}
\alias{seatrackgps-package}
\title{seatrackgps: Utility functions for working with GPS within the seatrack project.}
\description{
Implements functions to assist reading and organising GPS data within the seatrack project.
}
\author{
\strong{Maintainer}: Julian Evans \email{julian.evans@nina.no}

}
\keyword{internal}
