% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seatrackRgps-package.R
\docType{package}
\name{seatrackRgps-package}
\alias{seatrackRgps}
\alias{seatrackRgps-package}
\title{seatrackRgps: Utility functions for working with GPS within the seatrack project.}
\description{
Implements functions to assist reading and organising GPS data within the seatrack project.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ninanor.github.io/seatrack-gps/}
}

}
\author{
\strong{Maintainer}: Julian Evans \email{julian.evans@nina.no}

}
\keyword{internal}
