% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_handling_functions.R
\name{read_pos_data}
\alias{read_pos_data}
\title{Get position data from a specific tag file}
\usage{
read_pos_data(tag_id, file_connection)
}
\arguments{
\item{tag_id}{The ID of the tag to retrieve position data for.}

\item{file_connection}{FILE CONNECTION}
}
\value{
A data frame containing the position data with a POSIXct date_time column
and position measurements.
}
\description{
Get position data from a specific tag file
}
\examples{
\dontrun{
get_pos_data("61029", "path/to/directory")
}
}
